+++ DataViews/SampleType2.cs
@@ -8,7 +8,9 @@ namespace DataViews
         [SdsMember(IsKey = true)]
         public DateTime Time { get; set; }
 
+        [SdsMember(Uom = "bar")]
         public double Pressure { get; set; }
+        [SdsMember(Uom = "degree Celsius")]
         public double AmbientTemperature { get; set; }
     }
 }