+++ HISTORY.md
@@ -1,5 +1,11 @@
 # Version History
 
+## 1.1.0 / 2021-05-19
+
+- Added UoM Columns
+- Added Summaries Columns
+- Added helper functions
+
 ## 1.0.0 / 2021-05-03
 
 - Initial release